롤링바 생성 함수 컴포넌트화
